Configurator Mode toggle improvement
The Configurator Mode toggle now also makes sure the edit mode is properly disabled when needed
